requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
